ENH: Addition of rotation of the permuted and bootstrapped matrix.
Spatiotemporal analysis of event-related fMRI data using partial
least squares - A.R. McIntosh, * W.K. Chau, and A.B. Protzner

During resampling with either permutations or bootstrapping,
there is a possibility of axis rotation (a change in the order of
extracted LVs) and reflection (a sign change in saliences) when
the SVD is performed on the resampled matrix. These arbitrary
reflections and rotations can be corrected for through Procrustes
rotation of the resampled SVD outcome to the SVD outcome of
the original data (Milan and Whittaker, 1995). For the permuta-
tion, the Procrustes rotation ensures that the assessment is done on
the exact same effect space for each iteration. For bootstrapping,
the Procrustes rotation reduces the estimation bias because of
reflection and rotation.